FEAT: add universal checkpoint utils
