+++ fengshen/utils/__init__.py
@@ -1,0 +1,2 @@
+from .universal_checkpoint import UniversalCheckpoint
+__all__ = ['UniversalCheckpoint']
